+++ backend/messaging/wizard_topic.go
@@ -1,49 +1,90 @@
 package messaging
 
-import "sync"
+import (
+	"context"
+	"errors"
+	"sync"
 
-// WizardTopic encapsulates state around a WizardBus topic and its
+	"github.com/sensu/sensu-go/types"
+)
+
+// wizardTopic encapsulates state around a WizardBus topic and its
 // consumer channel bindings.
-type WizardTopic struct {
+type wizardTopic struct {
+	id       string
+	bindings map[string]Subscriber
+	ring     types.Ring
 	sync.RWMutex
-	Bindings map[string]chan<- interface{}
 }
 
 // Send a message to all subscribers to this topic.
-func (wTopic *WizardTopic) Send(msg interface{}) {
+func (wTopic *wizardTopic) Send(msg interface{}) {
 	wTopic.RLock()
 	defer wTopic.RUnlock()
 
-	for _, ch := range wTopic.Bindings {
+	for _, subscriber := range wTopic.bindings {
 		select {
-		case ch <- msg:
+		case subscriber.Receiver() <- msg:
 		default:
 			continue
 		}
 	}
 }
 
-// Subscribe a channel, identified by a consumer name, to this topic.
-func (wTopic *WizardTopic) Subscribe(consumer string, channel chan<- interface{}) {
+// SendDirect sends a message directly to a subscriber of this topic.
+func (wTopic *wizardTopic) SendDirect(msg interface{}) error {
+	if wTopic.ring == nil {
+		return errors.New("no ring for topic: " + wTopic.id)
+	}
+
+	wTopic.RLock()
+	defer wTopic.RUnlock()
+
+	id, err := wTopic.ring.Next(context.Background())
+	if err != nil {
+		return err
+	}
+
+	wTopic.bindings[id].Receiver() <- msg
+
+	return nil
+}
+
+// Subscribe a Subscriber to this topic and receive a Subscription.
+func (wTopic *wizardTopic) Subscribe(id string, sub Subscriber) (Subscription, error) {
 	wTopic.Lock()
-	wTopic.Bindings[consumer] = channel
-	wTopic.Unlock()
+	defer wTopic.Unlock()
+
+	wTopic.bindings[id] = sub
+
+	if wTopic.ring != nil {
+		if err := wTopic.ring.Add(context.Background(), id); err != nil {
+			return Subscription{}, err
+		}
+	}
+
+	return Subscription{
+		id:     id,
+		cancel: wTopic.unsubscribe,
+	}, nil
 }
 
 // Unsubscribe a consumer from this topic.
-func (wTopic *WizardTopic) Unsubscribe(consumer string) {
+func (wTopic *wizardTopic) unsubscribe(id string) {
 	wTopic.Lock()
-
-	delete(wTopic.Bindings, consumer)
-
+	delete(wTopic.bindings, id)
+	if wTopic.ring != nil {
+		// TODO: What do we do in the case of this failing?
+		_ = wTopic.ring.Remove(context.Background(), id)
+	}
 	wTopic.Unlock()
 }
 
 // Close all WizardTopic bindings.
-func (wTopic *WizardTopic) Close() {
+func (wTopic *wizardTopic) Close() {
 	wTopic.Lock()
-	for consumer := range wTopic.Bindings {
-		delete(wTopic.Bindings, consumer)
+	for consumer := range wTopic.bindings {
+		delete(wTopic.bindings, consumer)
 	}
 	wTopic.Unlock()
 }